backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# OCR
pytesseract==0.3.13
Pillow==11.1.0
pdf2image==1.17.0

# AI Classification (LangChain + Gemini)
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-core==0.3.29

# Export
openpyxl==3.1.5
pandas==2.2.3

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
